MagnonPhononHybridization

Documentation for MagnonPhononHybridization.

MagnonPhononHybridization.DMHybridization

MagnonPhononHybridization.MPHMetric

MagnonPhononHybridization.MagnonPhononCoupled

QuantumLattices.DegreesOfFreedom.Hilbert

QuantumLattices.DegreesOfFreedom.Metric

QuantumLattices.DegreesOfFreedom.Table

SpinWaveTheory.LSWT

QuantumLattices.add!

QuantumLattices.expand

TightBindingApproximation.commutator

MagnonPhononHybridization.DMHybridization — Type

DMHybridization(id::Symbol, value, bondkind; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=true)

The DM Magnon-Phonon coupling term.

Type alias for Term{:DMHybridization, id, V, B, C<:TermCoupling, A<:TermAmplitude, M<:TermModulate}

MagnonPhononHybridization.MPHMetric — Type

MPHMetric <: Metric

The metric of the operator indices of a magnon-phonon coupled system.

MagnonPhononHybridization.MagnonPhononCoupled — Type

MagnonPhononCoupled <: TBAKind{:BdG}

Magnon-phonon coupled quantum lattice system.

QuantumLattices.DegreesOfFreedom.Hilbert — Method

Hilbert(hilbert::Hilbert{<:CompositeInternal{:⊕, <:Union{Tuple{Phonon, Spin}, Tuple{Spin, Phonon}}}}, magneticstructure::MagneticStructure) -> Hilbert
Hilbert(hilbert::Hilbert{<:CompositeInternal{:⊗, <:Union{Tuple{Phonon, Spin}, Tuple{Spin, Phonon}}}}, magneticstructure::MagneticStructure) -> Hilbert

Get the hilbert space after the Holstein-Primakoff transformation of a magnon-phonon coupled system.

QuantumLattices.DegreesOfFreedom.Metric — Method

Metric(::MagnonPhononCoupled, ::Hilbert{<:CompositeInternal{K, <:Union{Tuple{Phonon, Fock}, Tuple{Fock, Phonon}}} where K}) -> MPHMetric

Get the metric of a magnon-phonon coupled system.

QuantumLattices.DegreesOfFreedom.Table — Method

Table(hilbert::Hilbert{<:CompositeInternal{K, <:Union{Tuple{Phonon, Fock}, Tuple{Fock, Phonon}}} where K}, by::MPHMetric) -> Table

Get the index-sequence table of a magnon-phonon couple system after the Holstein-Primakoff transformation.

SpinWaveTheory.LSWT — Method

LSWT(
    lattice::Lattice,
    hilbert::Hilbert{<:CompositeInternal{K, <:Union{Tuple{Phonon, Spin}, Tuple{Spin, Phonon}}} where K},
    terms::Tuple{Vararg{Term}},
    magneticstructure::MagneticStructure;
    neighbors::Union{Nothing, Int, Neighbors}=nothing,
    boundary::Boundary=plain
)

Construct a LSWT for a magnon-phonon coupled system.

QuantumLattices.add! — Method

add!(dest::Matrix, mr::TBAMatrixRepresentation{<:LSWT}, m::Operator{<:Number, <:Tuple{CompositeIndex{<:Index{Int, <:PID}}, CompositeIndex{<:Index{Int, <:FID{:b}}}}}; atol=atol/5, kwargs...) -> typeof(dest)

Get the matrix representation of an operator and add it to destination.

QuantumLattices.expand — Method

expand(::Val{:DMHybridization}, dmp::Coupling{<:Number, Tuple{Index{Colon, PID{Colon}}, Index{Colon, SID{wildcard, Colon}}}}, bond::Bond, hilbert::Hilbert) -> DMPExpand

Expand the default DM magnon-phonon coupling on a given bond.

TightBindingApproximation.commutator — Method

commutator(::MagnonPhononCoupled, hilbert::Hilbert{<:CompositeInternal{K, <:Union{Tuple{Phonon, Fock}, Tuple{Fock, Phonon}}} where K}) -> Matrix

Get the commutation relation of the Holstein-Primakoff bosons and phonons.