MagnonPhononHybridization.MagnonPhonon — Type

const MagnonPhonon = Union{Tuple{Phonon, Fock{:b}}, Tuple{Fock{:b}, Phonon}}

Internal magnon-phonon space.

MagnonPhononHybridization.SpinPhonon — Type

SpinPhonon = Union{Tuple{Phonon, Spin}, Tuple{Spin, Phonon}}

Internal spin-phonon space.

MagnonPhononHybridization.DMHybridization — Type

DMHybridization(id::Symbol, value, bondkind; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Union{Function, Bool}=true)

The DM Magnon-Phonon coupling term.

Type alias for Term{:DMHybridization, id, V, B, C<:TermCoupling, A<:TermAmplitude}

MagnonPhononHybridization.MPHMetric — Type

MPHMetric <: Metric

The metric of the operator indices of a magnon-phonon coupled system.

MagnonPhononHybridization.MagnonPhononCoupled — Type

MagnonPhononCoupled <: TBAKind{:BdG}

Magnon-phonon coupled quantum lattice system.

QuantumLattices.DegreesOfFreedom.Hilbert — Method

Hilbert(hilbert::Hilbert{<:InternalSum{<:SpinPhonon}}, magneticstructure::MagneticStructure) -> Hilbert
Hilbert(hilbert::Hilbert{<:InternalProd{<:SpinPhonon}}, magneticstructure::MagneticStructure) -> Hilbert

Get the hilbert space after the Holstein-Primakoff transformation of a magnon-phonon coupled system.

QuantumLattices.DegreesOfFreedom.Metric — Method

Metric(::MagnonPhononCoupled, ::Hilbert{<:CompositeInternal{<:MagnonPhonon}}) -> MPHMetric

Get the metric of a magnon-phonon coupled system.

QuantumLattices.DegreesOfFreedom.Table — Method

Table(hilbert::Hilbert{<:CompositeInternal{<:MagnonPhonon}}, by::MPHMetric) -> Table

Get the index-sequence table of a magnon-phonon couple system after the Holstein-Primakoff transformation.

SpinWaveTheory.LSWT — Method

LSWT(
    lattice::Lattice,
    hilbert::Hilbert{<:CompositeInternal{<:SpinPhonon}},
    terms::OneOrMore{Term},
    magneticstructure::MagneticStructure;
    neighbors::Union{Int, Neighbors}=nneighbor(terms)
)

Construct a LSWT for a magnon-phonon coupled system.

QuantumLattices.add! — Method

add!(dest::OperatorSum, qf::Quadraticization{MagnonPhononCoupled}, m::Operator{<:Number, <:NTuple{2, CoordinatedIndex{<:Index{<:FockIndex{:b}}}}}; kwargs...) -> typeof(dest)
add!(dest::OperatorSum, qf::Quadraticization{MagnonPhononCoupled}, m::Operator{<:Number, <:Tuple{CoordinatedIndex{<:Index{<:PhononIndex}}, CoordinatedIndex{<:Index{<:FockIndex{:b}}}}}; kwargs...) -> typeof(dest)

Get the matrix representation of an operator and add it to destination.

QuantumLattices.expand — Method

expand(dmp::Coupling{<:Number, <:Pattern{<:NTuple{2, Colon}, <:InternalPattern{(2,), <:Tuple{PhononIndex{:u}, SpinIndex}}}}, ::Val{:DMHybridization}, bond::Bond, hilbert::Hilbert) -> DMPExpand

Expand the default DM magnon-phonon coupling on a given bond.

TightBindingApproximation.commutator — Method

commutator(::MagnonPhononCoupled, hilbert::Hilbert{<:CompositeInternal{<:MagnonPhonon}}) -> Matrix

Get the commutation relation of the Holstein-Primakoff bosons and phonons.